onFillRequest: skip empty domain names
At least on a Pixel device running Android 11, an empty entry was also
added to the list of domains, and if that got picked as "first", the
result was a weird bug since no domain was displayed yet all domains got
routed to an empty string. This change ignores all empty strings from
the set of domains parsed from the AutoFill data source.